Timer stopping condition implemented

+++ SA_main.cpp
@@ -35,7 +35,7 @@ int main(int argc, char** argv) {
                 ANNEALING_STEPS_PER_ITERATION = atol(optarg);
                 break;
             case 't':
-                TOLERANCE = std::stod(optarg);
+                TOLERANCE = atol(optarg);
                 break;
             case '?':
                 if (optopt == 'x' || optopt == 'n' || optopt == 'i' || optopt == 'j' || optopt == 't') {
